refactor: extract hook implementations to separate files per Sisyphus rules
- Move blacklist-guard implementation from index.ts to hook.ts
- Move subagent-blacklist-guard implementation from index.ts to hook.ts
- index.ts now only re-exports, following project conventions
- Complies with Sisyphus Rule 1: index.ts is an entry point, not a dumping ground
